api logic - in progress
