+++ controllers/wordle_controller.py
@@ -1,0 +1,7 @@
+from services.wordle_service import wordle_logic
+
+def wordle_handler():
+    wordle_logic()
+
+def hello():
+    return "Hello World!"
