+++ services/format_file.py
@@ -3,7 +3,7 @@ from os import write
 
  
 def load_words():
-    with open('words.json') as json_file:
+    with open('services/words.json') as json_file:
         return json.load(json_file)
 
 if __name__ == '__main__':
@@ -14,5 +14,5 @@ if __name__ == '__main__':
             five_letter_words[word] = 1
   
     json_object = json.dumps(five_letter_words)
-    with open("five_letter_words.json", "w") as outfile:
+    with open("services/five_letter_words.json", "w") as outfile:
         outfile.write(json_object)
